pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-link"
version = "0.1.0"
description = "OCR invoices and extract structured JSON"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "pymupdf>=1.24",
    "Pillow>=10.0",
    "pytesseract>=0.3.10",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
llm = ["openai>=1.40"]

[project.scripts]
invoice-link = "invoice_link.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
